Move files parsing into another file
