+++ tavern/_plugins/rest/request.py
@@ -1,12 +1,10 @@
 import contextlib
 import json
 import logging
-import mimetypes
-import os
 import warnings
 from contextlib import ExitStack
 from itertools import filterfalse, tee
-from typing import Mapping, MutableMapping, Optional, Union
+from typing import Mapping, MutableMapping, Optional
 from urllib.parse import quote_plus
 
 import requests
@@ -20,6 +18,7 @@ from tavern._core.extfunctions import update_from_ext
 from tavern._core.general import valid_http_methods
 from tavern._core.pytest.config import TestConfig
 from tavern._core.report import attach_yaml
+from tavern._plugins.rest.files import get_file_arguments, guess_filespec
 from tavern.request import BaseRequest
 
 logger = logging.getLogger(__name__)
@@ -326,114 +325,6 @@ def _read_expected_cookies(
     return deep_dict_merge(from_cookiejar, from_extra)
 
 
-def _read_filespec(filespec: Union[str, dict]):
-    """
-    Get configuration for uploading file
-
-    Can either be just a path to a file or a 'long' format including content type/encoding
-
-    Args:
-        filespec: Either a string with the path to a file or a dictionary with file_path and possible content_type and/or content_encoding
-
-    Returns:
-        (file path, content type, content encoding)
-    """
-    if isinstance(filespec, str):
-        return filespec, None, None
-    elif isinstance(filespec, dict):
-        return (
-            filespec.get("file_path"),
-            filespec.get("content_type"),
-            filespec.get("content_encoding"),
-        )
-    else:
-        # Could remove, also done in schema check
-        raise exceptions.BadSchemaError(
-            "File specification must be a path or a dictionary"
-        )
-
-
-def _get_file_arguments(
-    request_args: dict, stack: ExitStack, test_block_config: TestConfig
-) -> dict:
-    """Get corect arguments for anything that should be passed as a file to
-    requests
-
-    Args:
-        request_args: args passed to requests
-        test_block_config: config for test
-        stack: context stack to add file objects to so they're
-            closed correctly after use
-
-    Returns:
-        mapping of 'files' block to pass directly to requests
-    """
-
-    files_to_send = {}
-
-    for key, filespec in request_args.get("files", {}).items():
-        file_spec = guess_filespec(filespec, stack, test_block_config)
-
-        files_to_send[key] = tuple(file_spec)
-
-    if files_to_send:
-        return {"files": files_to_send}
-    else:
-        return {}
-
-
-def guess_filespec(
-    filespec: Union[str, dict], stack: ExitStack, test_block_config: TestConfig
-):
-    """tries to guess the content type and encoding from a file.
-
-    Args:
-        test_block_config: config for test/stage
-        stack: exit stack to add open files context to
-        filespec: a string path to a file or a dictionary of the file path, content type, and encoding.
-
-    Returns:
-        tuple: A tuple of either length 2 (filename and file object), 3 (as before, with ceontent type), or 4 (as before, with with content encoding)
-
-    Notes:
-        If a 4-tuple is returned, the last element is a dictionary of headers to send to requests, _not_ the raw encoding value.
-    """
-    if not mimetypes.inited:
-        mimetypes.init()
-
-    filepath, content_type, encoding = _read_filespec(filespec)
-
-    filepath = format_keys(filepath, test_block_config.variables)
-    filename = os.path.basename(filepath)
-
-    # a 2-tuple ('filename', fileobj)
-    file_spec = [
-        filename,
-        stack.enter_context(open(filepath, "rb")),
-    ]
-
-    # Try to guess as well, but don't override what the user specified
-    guessed_content_type, guessed_encoding = mimetypes.guess_type(filepath)
-    content_type = content_type or guessed_content_type
-    encoding = encoding or guessed_encoding
-
-    # If it doesn't have a mimetype, or can't guess it, don't
-    # send the content type for the file
-    if content_type:
-        # a 3-tuple ('filename', fileobj, 'content_type')
-        logger.debug("content_type for '%s' = '%s'", filename, content_type)
-        file_spec.append(content_type)
-        if encoding:
-            # or a 4-tuple ('filename', fileobj, 'content_type', custom_headers)
-            logger.debug("encoding for '%s' = '%s'", filename, encoding)
-            # encoding is None for no encoding or the name of the
-            # program used to encode (e.g. compress or gzip). The
-            # encoding is suitable for use as a Content-Encoding header.
-            file_spec.append({"Content-Encoding": encoding})
-
-    return file_spec
-
-
 class RestRequest(BaseRequest):
     optional_in_file = [
         "json",
@@ -531,7 +422,7 @@ class RestRequest(BaseRequest):
                     request_args.update(data=file)
                 else:
                     self._request_args.update(
-                        _get_file_arguments(request_args, stack, test_block_config)
+                        get_file_arguments(request_args, stack, test_block_config)
                     )
 
                 headers = self._request_args.get("headers", {})
